refactor: Create an useOnlyOfficeContext
